requirements.txt
aerich==0.6.1
aiosqlite==0.17.0
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
cryptography==36.0.1
databases==0.5.3
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.63.0
fastapi-cloud-drives==0.1.4
FastAPI-SQLAlchemy==0.2.1
google-api-core==1.31.4
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
greenlet==1.1.2
h11==0.12.0
httplib2==0.20.2
idna==3.3
iso8601==0.1.16
Mako==1.1.6
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.3
passlib==1.7.4
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyDrive==1.3.1
PyDrive2==1.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pypika-tortoise==0.1.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scrypt==0.8.19
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.13.6
tomlkit==0.8.0
tortoise-orm==0.17.8
typing-extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.7
uvicorn==0.13.4
